requirements.txt
flask==3.0.3
flask-socketio==5.4.1
websocket-client==1.8.0 
loguru==0.7.2
